DDCNL-9021 | moved /finished into its own controller & tests
